refactor(IT-Cotato#230): Product 관련 enum 파일들을 enums 하위 폴더로 이동
- ProductCategory
- ProductSortType
- ProductStatus
