+++ src/main/java/com/cotato/kampus/domain/product/domain/ProductDetails.java
@@ -3,7 +3,7 @@ package com.cotato.kampus.domain.product.domain;
 import java.time.LocalDateTime;
 import java.util.List;
 
-import com.cotato.kampus.domain.product.ProductStatus;
+import com.cotato.kampus.domain.product.enums.ProductStatus;
 
 public record ProductDetails(
 	Long productId,
