+++ src/main/java/com/cotato/kampus/domain/product/enums/ProductCategory.java
@@ -1,4 +1,4 @@
-package com.cotato.kampus.domain.product;
+package com.cotato.kampus.domain.product.enums;
 
 public enum ProductCategory {
 	BEAUTY, CLOTHING, ELECTRONICS, FURNITURE, HOME_APPLIANCES, SPORTS, TOYS, ETC
