+++ src/main/java/com/cotato/kampus/domain/product/enums/ProductStatus.java
@@ -1,4 +1,4 @@
-package com.cotato.kampus.domain.product;
+package com.cotato.kampus.domain.product.enums;
 
 public enum ProductStatus {
 	ACTIVE, SOLD, RESERVED, DELETED
